myproject/requirements-myproject-frozen.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.2.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
deepdiff==8.5.0
defusedxml==0.7.1
distro==1.9.0
docx2txt==0.9
duckduckgo_search==8.0.4
durationpy==0.10
emoji==2.14.1
executing==2.2.0
fastapi==0.115.13
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
gradio==4.44.0
gradio_client==1.3.0
greenlet==3.2.3
groovy==0.1.2
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.0
humanfriendly==10.0
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.1.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.3.0
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.8.1
jupyter_nbextensions_configurator==0.6.4
jupyter_secrets_manager==0.4.0
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-anthropic==0.3.15
langchain-community==0.3.26
langchain-core==0.3.66
langchain-ibm==0.2.0
langchain-openai==0.3.25
langchain-tavily==0.1.5
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.1
langgraph-sdk==0.1.70
langsmith==0.4.2
lomond==0.3.3
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.1
mypy==1.16.1
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.22.0
openai==1.91.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orderly-set==5.4.1
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.3.8
posthog==5.4.0
primp==0.15.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pulsar-client==3.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.10.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruff==0.12.0
safehttpx==0.1.6
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.7.8
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.2
tomlkit==0.12.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unstructured==0.18.1
unstructured-client==0.25.9
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
